Put command processor into separate file
